+++ include/circt/Dialect/CMakeLists.txt
@@ -9,7 +9,6 @@
 add_subdirectory(AIG)
 add_subdirectory(Arc)
 add_subdirectory(Calyx)
-add_subdirectory(CalyxLib)
 add_subdirectory(Comb)
 add_subdirectory(DC)
 add_subdirectory(Debug)
@@ -27,6 +26,7 @@ add_subdirectory(LTL)
 add_subdirectory(Moore)
 add_subdirectory(MSFT)
 add_subdirectory(OM)
+add_subdirectory(OpLib)
 add_subdirectory(Pipeline)
 add_subdirectory(Ibis)
 add_subdirectory(Seq)
