+++ include/circt/Dialect/OpLib/OpLibDialect.h
@@ -1,4 +1,4 @@
-//===- CalyxLibDialect.h - CalyxLib Definitions -----------------*- C++ -*-===//
+//===- OpLibDialect.h - OpLib Definitions -----------------------*- C++ -*-===//
 //
 // Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
 // See https://llvm.org/LICENSE.txt for license information.
@@ -6,17 +6,17 @@
 //
 //===----------------------------------------------------------------------===//
 //
-// This file defines the CalyxLib CIRCT dialect.
+// This file defines the OpLib CIRCT dialect.
 //
 //===----------------------------------------------------------------------===//
 
-#ifndef CIRCT_DIALECT_CALYXLIB_CALYXLIBDIALECT_H
-#define CIRCT_DIALECT_CALYXLIB_CALYXLIBDIALECT_H
+#ifndef CIRCT_DIALECT_OPLIB_OPLIBDIALECT_H
+#define CIRCT_DIALECT_OPLIB_OPLIBDIALECT_H
 
 #include "circt/Support/LLVM.h"
 #include "mlir/IR/BuiltinOps.h"
 #include "mlir/IR/Dialect.h"
 
-#include "circt/Dialect/CalyxLib/CalyxLibDialect.h.inc"
+#include "circt/Dialect/OpLib/OpLibDialect.h.inc"
 
-#endif // CIRCT_DIALECT_CALYXLIB_CALYXLIBDIALECT_H
+#endif // CIRCT_DIALECT_OPLIB_OPLIBDIALECT_H
