+++ lib/Dialect/Calyx/CalyxOps.cpp
@@ -1671,12 +1671,14 @@ static LogicalResult
 verifyPrimitiveOpType(PrimitiveOp instance,
                       hw::HWModuleExternOp referencedPrimitive) {
   auto module = instance->getParentOfType<ModuleOp>();
-  StringRef entryPointName =
-      module->getAttrOfType<StringAttr>("calyx.entrypoint");
-  if (instance.getPrimitiveName() == entryPointName)
-    return instance.emitOpError()
-           << "cannot reference the entry-point component: '" << entryPointName
-           << "'.";
+  if (module->hasAttrOfType<StringAttr>("calyx.entrypoint")) {
+    StringRef entryPointName =
+        module->getAttrOfType<StringAttr>("calyx.entrypoint");
+    if (instance.getPrimitiveName() == entryPointName)
+      return instance.emitOpError()
+             << "cannot reference the entry-point component: '" << entryPointName
+             << "'.";
+  }
 
   // Verify the instance result ports with those of its referenced component.
   auto primitivePorts = referencedPrimitive.getPortList();
